+++ pythermalcomfort/models/set_tmp.py
@@ -93,13 +93,13 @@ def set_tmp(
         print(result.set)  # [24.3, 24.3]
 
     """
-    tdb = np.array(tdb)
-    tr = np.array(tr)
-    v = np.array(v)
-    rh = np.array(rh)
-    met = np.array(met)
-    clo = np.array(clo)
-    wme = np.array(wme)
+    tdb = np.asarray(tdb)
+    tr = np.asarray(tr)
+    v = np.asarray(v)
+    rh = np.asarray(rh)
+    met = np.asarray(met)
+    clo = np.asarray(clo)
+    wme = np.asarray(wme)
 
     # Validate inputs using the SetTmpInputs class
     SETInputs(