+++ pythermalcomfort/models/two_nodes_gagge.py
@@ -115,14 +115,14 @@ def two_nodes_gagge(
         w_max=w_max,
     )
 
-    tdb = np.array(tdb)
-    tr = np.array(tr)
-    v = np.array(v)
-    rh = np.array(rh)
-    met = np.array(met)
-    clo = np.array(clo)
-    wme = np.array(wme)
-    position = np.array(position)
+    tdb = np.asarray(tdb)
+    tr = np.asarray(tr)
+    v = np.asarray(v)
+    rh = np.asarray(rh)
+    met = np.asarray(met)
+    clo = np.asarray(clo)
+    wme = np.asarray(wme)
+    position = np.asarray(position)
 
     vapor_pressure = rh * p_sat_torr(tdb) / 100
 
@@ -516,11 +516,16 @@ def _gagge_two_nodes_optimized(
     elif t_body >= tbm_h:
         t_sens = w_max * 4.7 + 0.4685 * (t_body - tbm_h)
 
-    disc = (
-        4.7 * (e_rsw - e_comfort) / (e_max * w_max - e_comfort - e_diff)
-    )  # predicted thermal discomfort
+    if t_sens > 0 > e_max * w_max - e_comfort - e_diff:
+        disc = 6.0
+    else:
+        disc = (
+            4.7 * (e_rsw - e_comfort) / (e_max * w_max - e_comfort - e_diff)
+        )  # predicted thermal discomfort
     if disc <= 0:
         disc = t_sens
+    if disc > 6:
+        disc = 6.0
 
     # PMV Gagge
     pmv_gagge = (0.303 * math.exp(-0.036 * m) + 0.028) * (e_req - e_comfort - e_diff)